Added analytics logs

+++ Shared/src/main/java/com/lenta/shared/analytics/FmpAnalytics.kt
@@ -12,29 +12,49 @@ import kotlin.reflect.KFunction1
 
 class FmpAnalytics @Inject constructor(val hyperHive: HyperHive, val logDao: LogDao) : IAnalytics {
 
+    private var isEnabledLogs = false
+
+
+    override fun enableLogs(enable: Boolean) {
+        this.isEnabledLogs = enable
+    }
+
+
     override fun init() {
         hyperHive.loggingAPI.initAutoSendingLogs()
     }
 
 
     override fun logTrace(tag: String?, message: String) {
+        if (!isEnabledLogs) {
+            return
+        }
         logDao.insert(LogMessage(Date(), InfoLevel.INFO, message))
         enableLogsFuncDisableLogs(tag, message, hyperHive.loggingAPI::logTrace)
     }
 
 
     override fun logError(tag: String?, message: String) {
+        if (!isEnabledLogs) {
+            return
+        }
         logDao.insert(LogMessage(Date(), InfoLevel.ERROR, message))
         enableLogsFuncDisableLogs(tag, message, hyperHive.loggingAPI::logWarning)
     }
 
     override fun logFatal(tag: String?, message: String) {
+        if (!isEnabledLogs) {
+            return
+        }
         logDao.insert(LogMessage(Date(), InfoLevel.FATAL, message))
         enableLogsFuncDisableLogs(tag, message, hyperHive.loggingAPI::logFatal)
     }
 
 
     override fun cleanLogs() {
+        if (!isEnabledLogs) {
+            return
+        }
         hyperHive.databaseAPI.query(hyperHive.loggingAPI.loggingBasePath, "DELETE FROM hyperhive_journal").execute()
     }
 
@@ -45,16 +65,16 @@ class FmpAnalytics @Inject constructor(val hyperHive: HyperHive, val logDao: Log
 
     private fun enableLogsFuncDisableLogs(tag: String?, message: String, funcSendLogs: KFunction1<@ParameterName(name = "message") String, Unit>) {
         Logg.d { "logTrace: $tag, message: $message" }
-        enableLogs()
+        enableFmpLogs()
         funcSendLogs("${tag ?: ""}\n$message")
-        disableLogs()
+        disableFmpLogs()
     }
 
-    private fun enableLogs() {
+    private fun enableFmpLogs() {
         hyperHive.loggingAPI.setLogLevel(4)
     }
 
-    private fun disableLogs() {
+    private fun disableFmpLogs() {
         hyperHive.loggingAPI.setLogLevel(10)
     }
